Refactor(environment): modify workspace limits and update observation space
This commit updates the workspace limits slightly and adds a 5mm buffer
for the random coordinate choice, ensuring that point is withing the
workspace limits.

The set_start_method for the simulation environment was also changed to
make sure the desired random starting point perfectly matches the actual
position where the robot starts. This change aligns the robot
coordinates with the world coordinates in the simulation

Finally observation space was drastically changed. The joint positions
were dropped entirely and replaced by the Relative Vector space. All
values were also normalized in rande [-1, 1] to ensure a smooth gradient
flow for the RL models